refactor: Guess 에 있는 테스트 및 검증 책임 Word 로 이전
